[#36] feat: 테스트용으로 데이터를 리셋하는 기능 추가
