+++ backend/src/main/java/com/codesquad/baseball/controller/GameController.java
@@ -16,21 +16,25 @@ public class GameController {
         this.gameService = gameService;
     }
 
+    @CrossOrigin
     @GetMapping
     public GamesDTO showGames() {
         return gameService.showGames();
     }
 
+    @CrossOrigin
     @PatchMapping("/{gameId}")
     public void joinIn(@PathVariable("gameId") int gameId) {
         gameService.joinIn(gameId);
     }
 
+    @CrossOrigin
     @GetMapping("/{gameId}")
     public GameDetailDTO showGameDetail(@PathVariable("gameId") int gameId) {
         return gameService.gameDetail(gameId);
     }
 
+    @CrossOrigin
     @PostMapping("/{gameId}/pitch")
     public PitchDTO pitch(@PathVariable("gameId") int gameId) {
         return gameService.doPitch(gameId);
